src/requirements.txt
    azureml-mlflow
    scikit-learn
    pandas
    mlflow
    python-dateutil
    azure-identity
    azure-keyvault-secrets